// ClientPortal/src/models/Product.ts
import { OthersCodesInfo } from './OthersCodesInfo';
import { Price } from './Price';
import { ShortageInfo } from './ShortageInfo';

export interface ProductDetails {
    id: string;
    scrapingOriginIdentifier: string;
    scrapingOrigin: string;
    lastUpdate: string;
    originalName: string;
    name: string;
    country?: string;
    countryCode: string;
    countryName: string;
    productCode: string;
    atc: string;
    activeSubstances: string[];
    pharmaceuticalForm?: string;
    administrationRoute?: string;
    strength?: string;
    package?: string;
    isAuthorised: boolean;
    isMarketed?: boolean;
    isParallelImport?: boolean;
    statusNotes?: string;
    maHolder?: string;
    manufacturer?: string;
    isPsychotropic?: boolean;
    isGeneric?: boolean;
    documents?: { [key: string]: string };
    codes?: { [key: string]: string };
    //codes?: OthersCodesInfo;
    prices?: Price[];
    otherFields?: { [key: string]: string };
    strengthDetailed?: string;
    packageDetailed?: string;
    pharmaceuticalFormCategories?: string[];
    administrationCategories?: string[];
    autoName?: string;
    autoATC?: string;
    autoActiveSubstances?: string[];
    shortageInfo?: ShortageInfo;
    isBiological?: boolean;
    isAdditionalMonitoring?: boolean;
    isPrescription?: boolean;
    isHospitalar?: boolean;
    precautionsForStorage?: string;
}

export interface Product {
    data: {
        id: string;
        scrapingOriginIdentifier: string;
        scrapingOrigin: string;
        lastUpdate: string;
        originalName: string;
        name: string;
        country?: string;
        countryCode: string;
        countryName: string;
        productCode: string;
        atc: string;
        activeSubstances: string[];
        pharmaceuticalForm?: string;
        administrationRoute?: string;
        strength?: string;
        package?: string;
        isAuthorised: boolean;
        isMarketed?: boolean;
        isParallelImport?: boolean;
        statusNotes?: string;
        maHolder?: string;
        manufacturer?: string;
        isPsychotropic?: boolean;
        isGeneric?: boolean;
        documents?: { [key: string]: string };
        codes?: { [key: string]: string };
        //codes?: OthersCodesInfo;
        prices?: Price[];
        otherFields?: { [key: string]: string };
        strengthDetailed?: string;
        packageDetailed?: string;
        pharmaceuticalFormCategories?: string[];
        administrationCategories?: string[];
        autoName?: string;
        autoATC?: string;
        autoActiveSubstances?: string[];
        shortageInfo?: ShortageInfo;
        isBiological?: boolean;
        isAdditionalMonitoring?: boolean;
        isPrescription?: boolean;
        isHospitalar?: boolean;
        precautionsForStorage?: string;
    };
}